feat!: replace ServiceProvider with Scoper
The ServiceProvider interface was pairing the Resolver interface with
the ability to obtain Scopes which created a cycle: Scopes can create
ServiceProviders, ServiceProviders can create scopes, etc. This is fine,
but it implies, or at least raises the question whether Scopes obtained
from other Scopes are related. They're not related, so while it may be
slightly more ergonomic to take a single value that can be used as a
Resolver and to obtain new Scopes, it doesn't seem worth the ambiguity.
Reducing a Scope to something that provdes Resolvers and can be Closed
breaks the cycle, and the new Scoper interface allows the implmentation
of the Scope provider to remain hidden.