chore: remove ineffectual sentinel error
The `errInvalidResolution` is a vestige of when `ErrInvalidResolution`
and `InvalidResolution` were exported. Originally,
`InvalidResolution` provided a strongly-typed way to interact with the
failed resolution request's types, and `ErrInvalidResolution` was a
sentinel for the entire category of error. The types were removed from
the exported interface because there's no obvious usecase for
interacting with a failure resolution _programmatically_ -- if a type
doesn't implement an interface it seems like a code change is
inevitable. The only use for the strongly-typed error now is to control
the error message so sentinel is no longer useful.